Move BadgeList props to types file
